+++ EssentialFeed/EssentialFeed/FeedCache/FeedCachePolicy.swift
@@ -1,0 +1,23 @@
+//
+//  FeedCachePolicy.swift
+//  EssentialFeed
+//
+//  Created by Breno Valadão on 16/01/22.
+//
+
+import Foundation
+
+internal final class FeedCachePolicy {
+    private static let calendar = Calendar(identifier: .gregorian)
+
+    private static var maxCacheAgeInDays: Int { 7 }
+        
+    private init() {}
+    
+    static func validate(_ timestamp: Date, against date: Date) -> Bool {
+        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else {
+            return false
+        }
+        return date < maxCacheAge
+    }
+}
